src/requirements.txt
djangorestframework
Django
psycopg2
psycopg2-binary
pillow
tensorflow
sorl-thumbnail
django-resized
Gunicorn
djangorestframework-jwt
Markdown
Django-Filter
wikipedia-api
